Create Tests program using Catch2 (WIP)
The only test which has been partially ported to the Catch2 framework is
the CCZ4 test and this just evaluates the current CCZ4 RHS in a box with
no comparison (i.e. it always passes).

Since we only have one program now, there is also now a Common directory
for headers shared between tests.