Build: Refactor makefiles
This adds a base Makefile that should be common to GRTeclyn examples
(once there is more than one) and eventually the tests too.

There is also a new amrex-objects target in the root makefile which
should fix #68 as it is a dependency of the examples targets although
this does rely on the examples keeping the same default configuration
(i.e. USE_MPI = TRUE, COMP = gnu).

I have added a base name to the BinaryBH example.