Tests: Add test for calculation of MatterWeyl4
This will also modify the current implementation of Weyl4::set_up and
Weyl4::calculate_mf to be inline (so header can be included with
MatterWeyl4 without multiple definitions). It also changes the
private members of Weyl4 to be protected, so that MatterWeyl4 can
access the derivatives and epsilon3_U_LL and EB fields calculated by
Weyl4.