Port fixed grids tagging (#85)
* Tagging: Port fixed grid tagging

Also change from chi tagging to fixed grid tagging in BinaryBH
example

BinaryBH: Remove fill patch

Fill patch not needed for fixed grid tagging

* Tagging: Address Linter report

Also reverts back to original chi tagging criterion for BinaryBHLevel
because I had changed it to fixed grid tagging for the weak scaling
tests. I guess it should be chi tagging by default.

* Tagging: Remove current cell from function signature

As per Miren's review of PR#85, I've removed passing in the current cell
to the compute function as it is not needed in the calculations.