Core: Refactor GRAMRLevel::errorEst
This function is a pure virtual one in AmrLevel but it has a lot of
redundant arguments which we don't use. For simplicity, it now calls two
virtual functions with fewer arguments.
* pre_tag_cells()
* tag_cells()
which can be implemented in the example's derived class. The former is
optional and the latter is mandatory (pure virtual).

This change brings us closer to how things were in GRChombo.