Tagging: Generalize tagging criteria
Rather than requiring the use of a tagging criterion which calculates a
value for each cell and then forcing the tagging of cells where the
value exceeds the regrid_threshold, we now allow for "taggers" which
explicitly tag cells as they wish. This is a bit more flexible and it
avoids having to use a large magic number (e.g. 100.0) to ensure cells
we want to be tagged are definitely tagged independent of the
regrid_threshold.