Fixed MPI bug in power spectrum
- Had found that MPI threw errors when I used std::vectors to hold
the power spectrum. This was likely due to the vector going out of
scope before that MPI rank was done writing to it, for instance.
- Ended up adapting Yoann's method to what I wanted to calculate.
I use a map object to hold a tuple containing the isotropic k,
integrated power and count for each integer bin between 0 and N/2.
- I then print out the normalised power for each of these integer bins.
- I've run this and confirmed that it prints the right number of bins
with about the right overall behavior, but need to do a precision check
against GRChombo still.