Port of statistics functionality
- Created a class to calculate and print all requested momenta of a
given field (currently hard-coded for tensors, will need to think about
how to do this with arbitrary fields as we need to use arrays to store
the statistics and these are not dynamically allocated...)
- Created sub-routines to find the statistic (at order higher than 1)
and add the statistic along with a header for that stat.
- The function should be written such that the user can request any
combination of 1 (mean), 2 (stdev), 3 (skew) and 4 (kurt) in params
and ONLY those statistics will be calculated (after stdev which is needed
for all higher order stats), and only the requested stats will be printed.
But I need to spot check this functionality still.
- Added the params option but this doesn't work yet, need to port how
to read in the array. Currently the option is hard-coded right before the
function call in extract().