Fixed MPI bug in tensor moment calculation
- Each MPI rank has its own copy of sum in the calculation of
these statistics. I implemented a ParallelAllReduce to sum across
all MPI ranks before normalisation and assignment to print vector,
since all sum variable are initialised to zero. I could see this
producing bugs in the future if this initialisation is removed, and
in that case just saving which MPI threads calculated anything (somehow)
and summing/printing on these could be better. But I did it this
way because I anticipated an issue where perhaps multiple MPI threads
do different parts of the sum (if the hs_x array ends up being based
on state, for some reason, this could happen) and in this case the
global MPI sum will fix this issue.
- Removed header from spectra (just makes visualisation kind of worse).