Added diagnostic variables for polarisation fields
- Added diagnostic variables for plus and cross polarisation fields
in the RF class and in variableSetUp() in ScalarFieldLevel
- Added a routine in RF class to derive these polarisation fields
and put them on a diagnostic MF, then called this in a new section
of derive() in ScalarFieldLevel.
- This code compiles, not sure if it runs.
- I just copied the code that used to do this in extract(), so the
mode functions here are calculated twice, once in derive() and once
in extract(). Will need to streamline this later.
- Also the function in derive() isn't actually called yet. But it
compiles so that's something.