Tests: Add puncture tracker test
This test sets 1 component of the shift to a constant value and then
tracks two "punctures" which should just move in the opposite direction
to the shift. This is checked for as we know how they should move.

There is puncture tagging in the test but it's not currently checked.

Unfortunately, for the moment, we do have a common
SimulationParameters.hpp file for all the tests.