Matter: Refactor EMTensor
Now the choice of which components are stored is decided by the template
parameter em_tensor_parts.

The choice is more limited as we now assume that if you want the
momentum density, you will also want the energy density and similarly if
you want the stress tensor, you want everything else too.