CCZ4: Remove enum from diff1 and diff2
- Adds diff1 and diff2 functions for a. a single variable, specified
by a given enum index, and b. for all variables where each derivative
is stored in a simple GpuArray.
- Altered CCZ4RHS::compute to take diff1 of all variables and take diff2
of only those variables needed for the RHS.
- Altered the call signature of compute_rhs to take in the new diff
types, and altered the body of this function to be compatible with the
new types.
- Altered CCZ4::Geometry::compute_ricci_z to take the new diff1 and
diff2 of chi and h.
- Altered the moving puncture gauge to take the new type of diff1
and removed the diff2 argument (as this is no longer known by the
compute_rhs function directly).
- Kept the original compute_ricci_z function so that Wel4 does not
  break, and made the new function overload for the new diff types.