Altered default constraint handling
My previous implementation of the Ham_abs and Mom_abs printing seems
to have been buggy. There was some sort of memory leak that got bad as
MPI turned on.

Here, I've gone back to using the default constraint handling in
ScalarFieldLevel.cpp, denoted by "constraints" rather than
"constraints_norm", and have simply gone into Constraints.impl.hpp
and made it so the abs versions of the constraints are printed,
instead of the raw versions.

This is NOT a long-term fix, as it means you can't print the raw
constraints without changing the hard coding, but it's good enough
for now.