Fixed bug when running at N=512?
Figured out that for running at N=512, for some reason the code wants
me to use std::pow instead of * to multiply values of the indices in
the denominator of region 2 of the polarisation tensor construction
function.

Not sure why std::pow works and * doesn't, so it's possible this isn't
the true bug ie. the code may break tomorrow for some other reason...
should ask Miren or Juliana about this.