Core: Remove MPI communicator from amrex:Initialize
This should remove the need for an additional macro to wrap
up MPI_COMM_WORLD as we don't use our own communicator
anyway. For the main GRTeclyn amrex::Initialize, I added a
preprocessor guard for non-MPI builds so that the definition
of MPI_COMM_WORLD in the amrex namespace is used instead.