Fixed PS bug
1. Added the spectrum printing where we take the mean of the upper and
lower k on the interval, to interpolate onto the diagonal. This was a known
bug from GRChombo, it just hadn't been fixed here yet.
2. Removed factor of 2 in random magnitude assignment to get the magnitudes
right compared to Mathematica.
3. Fixed a bug where H0 in RandomFields was still being calculated according
to the m^2 phi^2 model, fixed by adding class-based H0 parameter and reading
in H0 from the grid (note: this will only be valid as long as scalar perts
aren't added on the first slice).