Bug fix in calculation of H0 in RandomFields
The previous way of finding H0 for the RandomFields class, by reading
K from the first point in the grid, wasn't safe for large N (256), and
caused a crash via NaN in hij(x).

I fixed this by copying the method from postSpecificTimeStep, where
H0 is found by finding the mean of k through the box using the built-in
sum function.