Made random draws MPI safe
Turns out the snippet of code Weiqun suggested doesn't work with
this implimentation as the underlying LoopOnCpu loop runs i first, then
j, then k. This loops over the grid towards the MPI barriers, not
parallel to the MPI barriers, which kept us from aligning the random
draws with each other across different MPI slicings.

I've fixed this bug by hard implimenting the box loop which draws the
random numbers from the generator, and swapping the order of loop so that
we loop over k first, then j, then i.

This produces a vector of random numbers which matched in coordinate
space when tested on 2 and 4 ranks with N=32. The spectra produced by these
runs also matched.