CI: Add regression test for A100s  (#93)
* CI: Fix the location of AMReX in an environment variable

This will add an environment variable AMREX_HOME for AMReX to be
installed so the Plotfile tools can be accessed later on without
having to navigate the directory structure.

* CI: Add fcompare for regression testing

This compares the output of params_test.ini with a pre-existing
file in the .github/workflows/data directory

* CI: Run fcompare to do the regression test

* CI: Remove some debug commands and add fcompare as make target

I removed the ls to check the output of the test run. Also, instead
of building all AMReX plot tools I've added project=fcompare to
the make command so only fcompare is built.

* CI: Add separate job for building tests on PVCs

- Start another job called csd3-dawn
- Only set up the environment modules for now

* CI: Build fcompare on csd3-dawn test job

- Add line to start an interactive session on Dawn
- Build fcompare within interactive session
- Also fix a typo from previous commit, should be projects=...

* CI: Only load Dawn modules when needed

- Dawn modules are now only loaded at the start of an
interactive job

* CI: Replace salloc with srun on csd3-dawn

- Interactive jobs via salloc don't seem to be allowed so
switching to srun instead

* CI: Make and run tests on Dawn pipeline

* CI: Remove Dawn pipeline and turn on HDF5 for the tests

* CI: Change version of HDF5 loaded

Switch to version not compiled with MPI support because the tests don't
use MPI, also turn off HDF5 for the BinaryBH regression test.

* CI: Add file lock to prevent simultaneous slurm jobs

Only one job can be submitted to the interactive queue at once, so
I've added a file lock on a dummy file, file.lock, such that the
srun commands are submitted to the queue one at a time

* CI: Add after_script to job to remove locked file

This is in case the srun command within flock fails. Also change
the location of the locked file to be common to both times flock
is called (so I can delete it).

* CI: Switch to MPI builds of HDF5

This will also:
  - Build tests with MPI
  - Fix definition of AMREX_HOME
  - Clean up flock and LOCKED_FILE variable

* Tests: Change GNUmakefile to use SLURM_NTASKS for MPI jobs

This commit will:
  - Change run to use the number of available MPI ranks as per
    SLURM_NTASKS. If SLURM_NTASKS is not defined, then leave
    the number of tasks as 2.

* CI: Change the number of tasks requested by Slurm to 2

* CI: Change srun to salloc for make

* CI: Change the number of tasks requested by Slurm to 2

* CI: Add loop to check for the existence of the plotfile

* CI: Add CCACHE to build config

* CI: Build the particle compare tool

* CI: Add particle regression test

* CI: Split common environment variables into separate file

* CI: Split common steps into separate file

This will also introduce a default before_script
and after_script

* CI: Break up test stage into a separate build stage

* CI: Move built items to build folder

To share built items between stages, I've moved them
to a folder outside of the usual GitLab CI location.
This is because the execs are too large to use
GitLab artifacts.

* CI: Move while loop to shell script

This will also add a formatting check for shell scripts that is
consistent with the GitLab style guide as a pre-commit hook

* CI: Move GitLab CI yaml files to own directory

* CI: Make the location of STORE unique to each pipeline

The directory name now contains the CI_PIPELINE_ID variable so
subsequent jobs won't overwrite the contents of the directory.
I also added an extra line to the test stage to remove the
directory afterwards.

* CI: Rename STORE to BINARY_DIR

I believe BINARY_DIR is a more descriptive name for
the location of the executables produced in the test
stage. I've also placed the removal of this directory
in the specific after_script of the test stage.

* CI: Refactor while loop and BINARY_DIR

This will move the definition of BINARY_DIR to the before_script
so there is no need for a dotenv artifact. Also I changed the
ordering of the while loop so it is hopefully clearer when things
go wrong with the filesystem.

* CI: Store outputs in RDS dp002 directory

* CI: Fix loop formatting

The output directory is not being deleted because the if
statement is never evaluated

* CI: Refactor environment variables

The directory storing the binaries and the outputs weren't being
deleted because the after_script doesn't have access to
environment variables defined in the before_script and in the
script

* CI: Move test build into build stage