Remove algebraic constraint enforcement in specificEvalRHS
This commit addresses Issue #101 by removing the algebraic constraint
enforcement functions i.e. TraceARemoval() and PositiveChiAndLapse()
from the specificEvalRHS function in the BinaryBH example.

By systematically removing each call to these functions in the BinaryBH
example, I determined that the call to TraceARemoval() and the calls
to both TraceARemoval() and PositiveChiAndLapse() in _either_ specificEvalRHS
or specificAdvance are required in order to prevent crash, up to 6 time
steps on the coarsest level, using the model given by params.txt.