KleinGordon: Refactor model classes
This will remove the MultiFab calculations from being class
member functions, which didn't work on GPUs. It will also
redefine the calculations as templated functions over the
model class.

Each model has also been rewritten so that their constructor
doesn't take any arguments (rather these are taken from the
ParmParse table allowing either of the models to be constructed
interchangably. (Otherwise the compiler will complain that
there is a mismatch in the number of arguments.)