First commit to add read in from STOIIC output
This starts off my effort to read in scalar perturbations from Yoann's
STOIIC pipeline. The alterations so far are:
1. Add a dparams.txt file with which to play with (should give a
   reasonable model).
2. Change SimulationParameters names to reflect some new parameter names
   in dparams.txt.
3. Made bug fixes from waaaay long ago, including adding type in first
   parameter of RandomFields::make_random_draws function and fixed
initialisation of bin_number in the RandomFields::params_t object (now
throws an error once it must be used instead of defaulting to
N_readin/2).
4. Took Mp rescaling out of InitialBackgroundData::compute, as this is
   taken into account by all of Yoann's initial values already.

This needs to be tested, but does compile at least.