SmallDataIO: Refactor reading of columns by name
This will remove the call to std::find on each line and replaced it
with a truth table instead where the columns selected by the user
have been marked as true.

This will also replace the vector containing the column numbers with
a std::map so the data will be returned with the same order as
requested by the user.