(BF) A. Moved invert index into get_kmag and B. add tensor init flag
A.
While ensuring I am directly reading in Yoann's ICs, I noticed that I
was not inverting the indices within the stoiic read in code, as I had
moved the calculation of kmag into a separate get_kmag method but had
not also moved the inverting of the indices, which meant I had to invert
the indices separately before the get_kmag method was called, or else I
would end up with kmags outside the domain, e.g. at (32, 63, 63).

To fix this, I moved the inverting of the indices into get_kmag and
removed them whenever they were called immediately before get_kmag. Note
I did NOT remove them from the tensor polarisation function or from the
Nyquist condition function, as these use the invert_index method in a
separate manner.

I also made get_kmag take an IntVect, because I always have an IntVect
around when I need to call this function and it's eaiser to read this
way.

I have also added print statements to the bit where the scalars are read
in from stoiic in params, in order to spot check that this read in is
correct. Confirmed the spot check with 1 resource and MPI/OpenMP
parallelism.

B.
I also added a flag to params that can be used to turn off the tensors,
as the next test Paul thinks we should do is to run with just tensors
and try to extract the second-order GWs with Charlotte's method.