CI: Add check-changes workflow to limit job execution
This will limit when these jobs are triggered by checking which paths
have been changed and they won't run e.g. if the README changes. This
has only been added to the workflows which actually build some code as
these take the most time.

The pull_request trigger has also been enabled to ensure that these
workflows are run in a PR on the final commit even if they're not
triggered on push.

Co-authored-by: Juliana Kwan <julianakwan123@gmail.com>