Geometry: Add support for negative extents
This removes all of our custom N* and L* parameters and instead moves to
relying on AMReX's geometry.prob_lo, geometry.prob_hi and amr.n_cell
parameters.  Note that we assume that dx is the same in every direction
in some parts of the code so a check for that has been added.

The CI will almost certainly fail since I have updated the regression
test so:

[skip ci]