Various cleaning and bug fixes
1. Removed calculation and printing of chi and phi
variances to the data file means_vars.dat, as we now
have a more precise method of capturing excitations
in the scalar section (statistics on the gauge-invariant
R field), and these statistics had been buggy (printing
nans) for a while.

2. Made read in of tensor initial conditions
conditional on tensors being used, to cut down on
time and memory load for runs that only initialise
the scalar sector.

3. Added print statements to denote when the scalar
power spectrum is being read in and when it is
being used to construct initial conditions, as these
are the heaviest parts of the IC routine for high-N
codes and this can help with timing.

4. Lowered the automatic location of the window function
by a factor of 2, to match with the current window
function used in the Mathematica solver, which I
lowered in order to cut down on high-f noise there.

5. Additionally made the filling of the initial hs,
As, hij and Aij MFs conditional on tensors being used
on the initial slice, to cut down on memory and time load.

6. Created error trap so that if tensors are not used
on the first slice, then their statistics are printed
as 0 on the first slice. Previously there was a bug here
where the skew/kurt of the pol fields on the first slice
returned nan, as I was calculating 0/0 essentially,
whereas these should just be 0. Now, anytime a
statistic is 0, the RandomField::find_field_moment_x
function just returns 0, rather than stat / vol.