Refactor Potential class and add USR potential
1. Change the construction of the Potential class,
so that the class contains individual member functions
that calculate each implimented potential, and the
member function called is chosen according to the
potential_type parameter helpfully supplied by STOIIC.

2. Moved the quadratic and monodromy potentials
to their own functions.

3. Added a new member function to calculate the USR
potential from the Prokopec paper, arXiv:2507.04114.

4. Tested that the quadratic potential returns the
expected value for V and dV on the initial slice.

5. Added a parameter to read in the potential
type from the params file, and read this in inside
of SimulationParameters.